feat: 무역 준비 데이터 흐름과 출발 어댑터 구현
- 무역 준비 Draft와 DraftStore 구현
- 마차 변경 시 동물 및 화물 선택 초기화
- 중복 고용이 불가능한 용병 선택 구조 추가
- TradePrepareViewData Builder 및 BuildContext 구현
- SaveData, Economy 가격 계산기, CaravanCalculator 연동
- 적재량, 슬롯, 식량, 용병 전투력 및 출발 조건 계산
- 화물 없음, 동물 혼합, 최대 견인 수, 슬롯 초과 검증 추가
- 미보유 마차·동물 및 유효하지 않은 용병 선택 차단
- M1 기준 화물 1종 제한 및 출발 전 판매 차단
- 실제 적재 화물과 선택 용병을 CaravanData로 변환
- TradePrepareFlowController와 TradeStartAdapter 구현
- 출발 실패 시 정산 Commit 데이터 롤백 처리
- 정산용 용병 비용 Commit 데이터 및 외부 전달 인터페이스 추가
- RouteData의 이전 비용 필드를 Legacy 호환 필드로 정리
- 식량 비용은 0, 용병 비용은 선택 용병 가격 합계로 변경
- 용병 전투력과 마차 내구도를 int로 통일

+++ Assets/99.Sandbox/_LJH/01.Script/Data/MercenaryData.cs
@@ -17,7 +17,7 @@ public class MercenaryData : ScriptableObject, IIdentifiableData
     [SerializeField] private long baseBuyPrice;
 
     [Header("Combat_Info")]
-    [SerializeField] private float combatCapability;
+    [SerializeField] private int combatCapability;
 
     [Header("Mercenary_Unlocked_Default")]
     [SerializeField] private bool unlockedByDefault;
@@ -31,7 +31,7 @@ public class MercenaryData : ScriptableObject, IIdentifiableData
     public string Description => description;
     public TradeItemRarity Rarity => rarity;
     public long BaseBuyPrice => baseBuyPrice;
-    public float CombatCapability => combatCapability;
+    public int CombatCapability => combatCapability;
     public bool UnlockedByDefault => unlockedByDefault;
     #endregion
 }
